Refactor profile router, add tests, fix warnings
- Add app/routers/profile.py with all profile routes
- Add app/core/dependencies.py with get_current_user
- Add ProfileService.edit_profile and ProfileRepository.edit_profile
- Add validation to ProfileCreateDTO using Field()
- Fix AthleteProfile.weekly_hours type int -> float
- Fix TemplateResponse deprecated signature across all files
- Fix exception_handler missing status_code
- Add tests: test_non_auth_user_error, test_edit_profile_success

+++ app/models/models.py
@@ -50,7 +50,7 @@ class AthleteProfile(SQLModel, table=True):
     weight_kg: Optional[float]
     current_ftp: Optional[int]
     limitations: Optional[str]
-    weekly_hours: Optional[int]  # Сколько времени есть на тренировки в неделю
+    weekly_hours: Optional[float]  # Сколько времени есть на тренировки в неделю
     gear: Optional[str]  # какое оборудование есть у пользователя(велосипед, станки..)
     environment_location: Optional[str]  # окружение, локация
     user: Optional['Users'] = Relationship(back_populates='athlete_profile')
